refactor(api): decompose strategy autopilot
