+++ apps/server/api/src/collections/agent-strategies/services/agent-strategy-autopilot.service.ts
@@ -1,316 +1,65 @@
-import { ActivitiesService } from '@api/collections/activities/services/activities.service';
-import { type AgentStrategyDocument } from '@api/collections/agent-strategies/schemas/agent-strategy.schema';
-import { type AgentStrategyOpportunityDocument } from '@api/collections/agent-strategies/schemas/agent-strategy-opportunity.schema';
+import type { AgentStrategyDocument } from '@api/collections/agent-strategies/schemas/agent-strategy.schema';
+import type { AgentStrategyOpportunityDocument } from '@api/collections/agent-strategies/schemas/agent-strategy-opportunity.schema';
 import { AgentStrategyReportType } from '@api/collections/agent-strategies/schemas/agent-strategy-policy.schema';
 import { AgentStrategiesService } from '@api/collections/agent-strategies/services/agent-strategies.service';
+import type {
+  AgentStrategyPerformanceSnapshot,
+  ExecuteRunResult,
+} from '@api/collections/agent-strategies/services/agent-strategy-autopilot.types';
+import { AgentStrategyAutopilotExecutionService } from '@api/collections/agent-strategies/services/agent-strategy-autopilot-execution.service';
+import { AgentStrategyAutopilotPerformanceService } from '@api/collections/agent-strategies/services/agent-strategy-autopilot-performance.service';
+import { AgentStrategyAutopilotPlanningService } from '@api/collections/agent-strategies/services/agent-strategy-autopilot-planning.service';
 import { AgentStrategyOpportunitiesService } from '@api/collections/agent-strategies/services/agent-strategy-opportunities.service';
-import { AgentStrategyReportsService } from '@api/collections/agent-strategies/services/agent-strategy-reports.service';
-import { BrandsService } from '@api/collections/brands/services/brands.service';
-import { type ContentDraftDocument } from '@api/collections/content-drafts/schemas/content-draft.schema';
-import { ContentDraftsService } from '@api/collections/content-drafts/services/content-drafts.service';
-import { ContentPerformanceService } from '@api/collections/content-performance/services/content-performance.service';
-import { CredentialsService } from '@api/collections/credentials/services/credentials.service';
-import { EvaluationsOperationsService } from '@api/collections/evaluations/services/evaluations-operations.service';
-import { OptimizersService } from '@api/collections/optimizers/services/optimizers.service';
-import { PostsService } from '@api/collections/posts/services/posts.service';
-import { TrendsService } from '@api/collections/trends/services/trends.service';
 import { NotFoundException } from '@api/helpers/exceptions/http/not-found.exception';
-import { BatchGenerationService } from '@api/services/batch-generation/batch-generation.service';
-import { ReviewBatchItemFormat } from '@api/services/batch-generation/constants/review-batch-item-format.constant';
-import { ContentGatewayService } from '@api/services/content-gateway/content-gateway.service';
-import {
-  ActivityEntityModel,
-  ActivityKey,
-  ActivitySource,
-  ContentDraftStatus,
-  ContentFormat,
-  IngredientCategory,
-  PostCategory,
-  PostStatus,
-} from '@genfeedai/enums';
-import { LoggerService } from '@libs/logger/logger.service';
 import { Injectable } from '@nestjs/common';
-import { PerformanceSummaryService } from '@server/collections/content-performance/services/performance-summary.service';
 
-export interface BudgetPacingState {
-  expectedSpendToDate: number;
-  monthBudget: number;
-  monthToDateCreditsUsed: number;
-  remainingDailyBudget: number;
-  remainingMonthlyBudget: number;
-  remainingWeeklyBudget: number;
-  reserveTrendBudgetRemaining: number;
-}
-
-export interface PublishGateResult {
-  decision: 'approved' | 'discard' | 'hold' | 'revise';
-  overallScore: number;
-  reasons: string[];
-  revisionInstructions: string[];
-  scoreBreakdown: Record<string, number>;
-}
-
-export interface AgentStrategyPerformanceSnapshot {
-  bestPlatformFormatPairs: Array<{
-    format: string;
-    platform: string;
-    score: number;
-  }>;
-  bestPostingWindows: string[];
-  clicks: number;
-  costPerVisit: number;
-  creditsSpent: number;
-  ctr: number;
-  generatedCount: number;
-  impressions: number;
-  publishedCount: number;
-  topHooks: string[];
-  topTopics: string[];
-  visits: number;
-}
-
-interface ExecuteRunResult {
-  contentGenerated: number;
-  creditsUsed: number;
-  summary: string;
-}
-
-interface OptimizerAnalysisResult {
-  breakdown?: {
-    clarity?: number;
-    engagement?: number;
-    platformOptimization?: number;
-    readability?: number;
-  };
-  metadata?: {
-    hasCallToAction?: boolean;
-  };
-  overallScore?: number;
-}
-
-interface ImageEvaluationScore {
-  overall?: number;
-}
-
-interface ImageEvaluationResult {
-  overallScore?: number;
-  scores?: {
-    brand?: ImageEvaluationScore;
-    engagement?: ImageEvaluationScore;
-    technical?: ImageEvaluationScore;
-  };
-}
-
-interface FinalizeOpportunityInput {
-  draft: ContentDraftDocument;
-  draftContent: string;
-  format: string;
-  gate: PublishGateResult;
-  opportunity: AgentStrategyOpportunityDocument;
-  organizationId: string;
-  platform: string;
-  strategy: AgentStrategyDocument;
-  userId: string;
-}
+export type {
+  AgentStrategyPerformanceSnapshot,
+  BudgetPacingState,
+  PublishGateResult,
+} from '@api/collections/agent-strategies/services/agent-strategy-autopilot.types';
 
 @Injectable()
 export class AgentStrategyAutopilotService {
-  private readonly defaultEventOpportunityCost = 12;
-  private readonly defaultImageOpportunityCost = 24;
-  private readonly defaultTextOpportunityCost = 10;
-
   constructor(
     private readonly agentStrategiesService: AgentStrategiesService,
     private readonly opportunitiesService: AgentStrategyOpportunitiesService,
-    private readonly reportsService: AgentStrategyReportsService,
-    private readonly activitiesService: ActivitiesService,
-    private readonly trendsService: TrendsService,
-    private readonly brandsService: BrandsService,
-    private readonly contentGatewayService: ContentGatewayService,
-    private readonly contentDraftsService: ContentDraftsService,
-    private readonly optimizersService: OptimizersService,
-    private readonly evaluationsOperationsService: EvaluationsOperationsService,
-    private readonly credentialsService: CredentialsService,
-    private readonly postsService: PostsService,
-    private readonly batchGenerationService: BatchGenerationService,
-    private readonly contentPerformanceService: ContentPerformanceService,
-    private readonly performanceSummaryService: PerformanceSummaryService,
-    private readonly logger: LoggerService,
+    private readonly performanceService: AgentStrategyAutopilotPerformanceService,
+    private readonly planningService: AgentStrategyAutopilotPlanningService,
+    private readonly executionService: AgentStrategyAutopilotExecutionService,
   ) {}
 
   async listStrategyOpportunities(
-    strategyId: string,
+    strategyIdValue: string,
     organizationId: string,
   ): Promise<AgentStrategyOpportunityDocument[]> {
-    const strategy = await this.requireStrategy(strategyId, organizationId);
-    await this.refreshOpportunities(strategy);
-    return this.opportunitiesService.listOpenByStrategy(
-      strategyId,
+    const strategy = await this.requireStrategy(
+      strategyIdValue,
+      organizationId,
+    );
+    return this.planningService.refreshOpportunities(strategy);
+  }
+
+  getPerformanceSnapshot(
+    strategyIdValue: string,
+    organizationId: string,
+  ): Promise<AgentStrategyPerformanceSnapshot> {
+    return this.performanceService.getPerformanceSnapshot(
+      strategyIdValue,
       organizationId,
     );
   }
 
-  async getPerformanceSnapshot(
-    strategyId: string,
-    organizationId: string,
-  ): Promise<AgentStrategyPerformanceSnapshot> {
-    const strategy = await this.requireStrategy(strategyId, organizationId);
-    const { periodEnd, periodStart } = this.resolveReportWindow('weekly');
-
-    const strategyBrandId = this.strategyBrandId(strategy);
-    const strategyOrganizationId = this.strategyOrganizationId(strategy);
-
-    const [drafts, opportunities, performance, summary] = await Promise.all([
-      this.contentDraftsService.find({
-        brandId: strategyBrandId ?? String(strategy.brand),
-        createdAt: { gte: periodStart, lte: periodEnd },
-        isDeleted: false,
-        organizationId: strategyOrganizationId,
-      }),
-      this.opportunitiesService.listByStrategy(strategyId, organizationId),
-      strategy.brand
-        ? this.contentPerformanceService.queryPerformance(
-            {
-              brand: String(strategyBrandId),
-              endDate: periodEnd.toISOString(),
-              limit: 250,
-              startDate: periodStart.toISOString(),
-            },
-            organizationId,
-          )
-        : [],
-      strategy.brand
-        ? this.performanceSummaryService
-            .getWeeklySummary(organizationId, String(strategyBrandId), {
-              endDate: periodEnd,
-              startDate: periodStart,
-            })
-            .catch(() => null)
-        : null,
-    ]);
-
-    const impressions = performance.reduce((sum, item) => sum + item.views, 0);
-    const clicks = performance.reduce((sum, item) => sum + item.clicks, 0);
-    const visits = clicks;
-    const generatedCount = drafts.length;
-    const publishedCount = drafts.filter(
-      (draft) =>
-        String(draft.status ?? '').toLowerCase() ===
-        ContentDraftStatus.PUBLISHED,
-    ).length;
-    const creditsSpent = opportunities
-      .filter((opportunity) => {
-        const createdAt = opportunity.createdAt;
-        return (
-          ['approved', 'published'].includes(opportunity.status) &&
-          createdAt instanceof Date &&
-          createdAt >= periodStart &&
-          createdAt <= periodEnd
-        );
-      })
-      .reduce((sum, opportunity) => sum + opportunity.estimatedCreditCost, 0);
-    const ctr =
-      impressions > 0 ? Number(((clicks / impressions) * 100).toFixed(2)) : 0;
-    const costPerVisit =
-      visits > 0 ? Number((creditsSpent / visits).toFixed(2)) : 0;
-
-    const topicCounts = new Map<string, number>();
-    for (const opportunity of opportunities) {
-      const createdAt = opportunity.createdAt;
-      if (
-        !(createdAt instanceof Date) ||
-        createdAt < periodStart ||
-        createdAt > periodEnd
-      ) {
-        continue;
-      }
-      topicCounts.set(
-        opportunity.topic,
-        (topicCounts.get(opportunity.topic) ?? 0) + 1,
-      );
-    }
-
-    const topTopics = [...topicCounts.entries()]
-      .sort((a, b) => b[1] - a[1])
-      .slice(0, 5)
-      .map(([topic]) => topic);
-
-    const pairScores = new Map<
-      string,
-      { format: string; platform: string; score: number }
-    >();
-    for (const item of performance) {
-      const key = `${item.platform}:${item.contentType}`;
-      const existing = pairScores.get(key) ?? {
-        format: String(item.contentType),
-        platform: String(item.platform),
-        score: 0,
-      };
-      existing.score += item.performanceScore ?? 0;
-      pairScores.set(key, existing);
-    }
-
-    return {
-      bestPlatformFormatPairs: [...pairScores.values()]
-        .sort((a, b) => b.score - a.score)
-        .slice(0, 5),
-      bestPostingWindows:
-        summary?.bestPostingTimes
-          ?.slice(0, 3)
-          .map((item) => `${item.hour}:00`) ?? [],
-      clicks,
-      costPerVisit,
-      creditsSpent,
-      ctr,
-      generatedCount,
-      impressions,
-      publishedCount,
-      topHooks: summary?.topHooks ?? [],
-      topTopics,
-      visits,
-    };
-  }
-
-  async generateStrategyReport(
-    strategyId: string,
+  generateStrategyReport(
+    strategyIdValue: string,
     organizationId: string,
     reportType: AgentStrategyReportType = 'daily',
   ) {
-    const strategy = await this.requireStrategy(strategyId, organizationId);
-    const snapshot = await this.getPerformanceSnapshot(
-      strategyId,
+    return this.performanceService.generateStrategyReport(
+      strategyIdValue,
       organizationId,
-    );
-    const { periodEnd, periodStart } = this.resolveReportWindow(reportType);
-
-    const allocationChanges = snapshot.bestPlatformFormatPairs
-      .slice(0, 2)
-      .map(
-        (pair) =>
-          `Bias next runs toward ${pair.platform}/${pair.format} based on current performance.`,
-      );
-
-    return this.reportsService.createReport({
-      allocationChanges,
-      bestPlatformFormatPairs: snapshot.bestPlatformFormatPairs,
-      bestPostingWindows: snapshot.bestPostingWindows,
-      brandId: this.strategyBrandId(strategy) ?? String(strategy.brand),
-      clicks: snapshot.clicks,
-      costPerVisit: snapshot.costPerVisit,
-      creditsSpent: snapshot.creditsSpent,
-      ctr: snapshot.ctr,
-      generatedCount: snapshot.generatedCount,
-      impressions: snapshot.impressions,
-      organizationId: this.strategyOrganizationId(strategy),
-      periodEnd,
-      periodStart,
-      publishedCount: snapshot.publishedCount,
       reportType,
-      strategyId: this.strategyId(strategy),
-      topHooks: snapshot.topHooks,
-      topTopics: snapshot.topTopics,
-      visits: snapshot.visits,
-    } as never);
+    );
   }
 
   async executeQueuedRun(input: {
@@ -324,11 +73,16 @@ export class AgentStrategyAutopilotService {
       input.strategyId,
       input.organizationId,
     );
-    const pacing = this.computeBudgetPacingState(strategy);
+    const pacing = this.planningService.computeBudgetPacingState(strategy);
 
     await this.opportunitiesService.expireStaleOpportunities(strategy);
-    const opportunities = await this.refreshOpportunities(strategy);
-    const selected = this.selectOpportunities(strategy, opportunities, pacing);
+    const opportunities =
+      await this.planningService.refreshOpportunities(strategy);
+    const selected = this.planningService.selectOpportunities(
+      strategy,
+      opportunities,
+      pacing,
+    );
 
     if (selected.length === 0) {
       return {
@@ -341,9 +95,8 @@ export class AgentStrategyAutopilotService {
 
     let generatedCount = 0;
     let creditsUsed = 0;
-
     for (const opportunity of selected) {
-      const result = await this.executeOpportunity(
+      const result = await this.executionService.executeOpportunity(
         strategy,
         opportunity,
         input.userId,
@@ -369,7 +122,7 @@ export class AgentStrategyAutopilotService {
       ?.dailyDigestEnabled
       ? 'daily'
       : 'weekly';
-    await this.generateStrategyReport(
+    await this.performanceService.generateStrategyReport(
       input.strategyId,
       input.organizationId,
       reportType,
@@ -383,1183 +136,16 @@ export class AgentStrategyAutopilotService {
   }
 
   private async requireStrategy(
-    strategyId: string,
+    strategyIdValue: string,
     organizationId: string,
   ): Promise<AgentStrategyDocument> {
     const strategy = await this.agentStrategiesService.findOneById(
-      strategyId,
+      strategyIdValue,
       organizationId,
     );
-
     if (!strategy) {
       throw new NotFoundException('Strategy');
     }
-
     return strategy;
-  }
-
-  private documentId(entity: unknown): string {
-    const record = entity as Record<string, unknown>;
-    return String(record.id);
-  }
-
-  private opportunityId(opportunity: AgentStrategyOpportunityDocument): string {
-    return this.documentId(opportunity);
-  }
-
-  private draftId(draft: ContentDraftDocument): string {
-    return this.documentId(draft);
-  }
-
-  private strategyId(strategy: AgentStrategyDocument): string {
-    return this.documentId(strategy);
-  }
-
-  private strategyOrganizationId(strategy: AgentStrategyDocument): string {
-    return (
-      this.asOptionalString(
-        (strategy as Record<string, unknown>).organizationId,
-      ) ?? strategy.organization
-    );
-  }
-
-  private strategyBrandId(strategy: AgentStrategyDocument): string | undefined {
-    return (
-      this.asOptionalString((strategy as Record<string, unknown>).brandId) ??
-      strategy.brand ??
-      undefined
-    );
-  }
-
-  private strategyPlatforms(strategy: AgentStrategyDocument): string[] {
-    return strategy.platforms?.length ? strategy.platforms : ['twitter'];
-  }
-
-  private strategySkillSlugs(
-    strategy: AgentStrategyDocument,
-    fallback: string[],
-  ): string[] {
-    return strategy.skillSlugs?.length ? strategy.skillSlugs : fallback;
-  }
-
-  private resolveOpportunityPlatform(
-    strategy: AgentStrategyDocument,
-    opportunity: AgentStrategyOpportunityDocument,
-  ): string {
-    return (
-      opportunity.platformCandidates[0] ?? this.strategyPlatforms(strategy)[0]
-    );
-  }
-
-  private normalizeOpportunitySourceType(
-    sourceType: AgentStrategyOpportunityDocument['sourceType'],
-  ): 'event' | 'evergreen' | 'trend' | undefined {
-    if (
-      sourceType === 'event' ||
-      sourceType === 'evergreen' ||
-      sourceType === 'trend'
-    ) {
-      return sourceType;
-    }
-
-    return undefined;
-  }
-
-  private asOptionalString(value: unknown): string | undefined {
-    return typeof value === 'string' && value.length > 0 ? value : undefined;
-  }
-
-  private computeBudgetPacingState(
-    strategy: AgentStrategyDocument,
-  ): BudgetPacingState {
-    const monthBudget = strategy.budgetPolicy?.monthlyCreditBudget ?? 500;
-    const now = new Date();
-    const daysInMonth = new Date(
-      now.getUTCFullYear(),
-      now.getUTCMonth() + 1,
-      0,
-    ).getUTCDate();
-    const currentDay = now.getUTCDate();
-    const expectedSpendToDate = Number(
-      ((monthBudget / daysInMonth) * currentDay).toFixed(2),
-    );
-
-    return {
-      expectedSpendToDate,
-      monthBudget,
-      monthToDateCreditsUsed: strategy.monthToDateCreditsUsed ?? 0,
-      remainingDailyBudget: Math.max(
-        0,
-        (strategy.dailyCreditBudget ?? 0) - (strategy.dailyCreditsUsed ?? 0),
-      ),
-      remainingMonthlyBudget: Math.max(
-        0,
-        monthBudget - (strategy.monthToDateCreditsUsed ?? 0),
-      ),
-      remainingWeeklyBudget: Math.max(
-        0,
-        (strategy.weeklyCreditBudget ?? 0) -
-          (strategy.creditsUsedThisWeek ?? 0),
-      ),
-      reserveTrendBudgetRemaining:
-        strategy.reserveTrendBudgetRemaining ??
-        strategy.budgetPolicy?.reserveTrendBudget ??
-        0,
-    };
-  }
-
-  private async refreshOpportunities(
-    strategy: AgentStrategyDocument,
-  ): Promise<AgentStrategyOpportunityDocument[]> {
-    const created: AgentStrategyOpportunityDocument[] = [];
-    const platforms = this.strategyPlatforms(strategy);
-    const defaultTopic =
-      strategy.topics?.[0] ?? strategy.label ?? 'General update';
-    const strategyBrandId = this.strategyBrandId(strategy);
-    const strategyId = this.strategyId(strategy);
-    const strategyOrganizationId = this.strategyOrganizationId(strategy);
-
-    if (strategy.opportunitySources?.trendWatchersEnabled && strategyBrandId) {
-      for (const platform of platforms.slice(0, 3)) {
-        const trends = await this.trendsService.getTrends(
-          strategyOrganizationId,
-          strategyBrandId,
-          platform,
-          { allowFetchIfMissing: false },
-        );
-
-        for (const trend of trends.slice(0, 3)) {
-          created.push(
-            await this.opportunitiesService.createIfMissing({
-              brandId: strategyBrandId,
-              decisionReason: 'Trend watcher matched a current platform trend.',
-              estimatedCreditCost: this.estimateOpportunityCost(
-                this.resolveFormatsForStrategy(strategy),
-              ),
-              expectedTrafficScore: Math.min(
-                100,
-                Math.round(trend.viralityScore ?? 0),
-              ),
-              expiresAt: trend.expiresAt
-                ? new Date(trend.expiresAt)
-                : undefined,
-              formatCandidates: this.resolveFormatsForStrategy(strategy),
-              metadata: {
-                platform,
-                trendId: String(trend.id),
-                viralityScore: trend.viralityScore ?? 0,
-              },
-              organizationId: strategyOrganizationId,
-              platformCandidates: [platform],
-              priorityScore: this.computePriorityScore(strategy, {
-                costEfficiency: 100 / this.defaultTextOpportunityCost,
-                expectedTraffic: trend.viralityScore ?? 0,
-                freshness: 90,
-                historicalConfidence: 50,
-                relevance: this.computeTopicRelevance(strategy, trend.topic),
-              }),
-              relevanceScore: this.computeTopicRelevance(strategy, trend.topic),
-              sourceRef: String(trend.id),
-              sourceType: 'trend',
-              strategyId,
-              topic: trend.topic,
-            }),
-          );
-        }
-      }
-    }
-
-    if (strategy.opportunitySources?.eventTriggersEnabled && strategyBrandId) {
-      const snapshot = await this.getPerformanceSnapshot(
-        strategyId,
-        strategyOrganizationId,
-      );
-      const topHook = snapshot.topHooks[0];
-
-      if (topHook) {
-        created.push(
-          await this.opportunitiesService.createIfMissing({
-            brandId: strategyBrandId,
-            decisionReason: 'Event trigger captured a high-performing hook.',
-            estimatedCreditCost: this.defaultEventOpportunityCost,
-            expectedTrafficScore: 75,
-            formatCandidates: ['text'],
-            metadata: { hook: topHook, trigger: 'high-performing-hook' },
-            organizationId: strategyOrganizationId,
-            platformCandidates: platforms,
-            priorityScore: this.computePriorityScore(strategy, {
-              costEfficiency: 100 / this.defaultEventOpportunityCost,
-              expectedTraffic: 75,
-              freshness: 70,
-              historicalConfidence: 80,
-              relevance: 80,
-            }),
-            relevanceScore: 80,
-            sourceRef: `event:hook:${topHook}`,
-            sourceType: 'event',
-            strategyId,
-            topic: topHook,
-          }),
-        );
-      }
-    }
-
-    if (strategy.opportunitySources?.evergreenCadenceEnabled) {
-      const recentPublishedCount = strategy.runHistory.filter((item) => {
-        if (!item.completedAt) {
-          return false;
-        }
-        const sevenDaysAgo = new Date();
-        sevenDaysAgo.setUTCDate(sevenDaysAgo.getUTCDate() - 7);
-        return item.completedAt >= sevenDaysAgo;
-      }).length;
-
-      if (recentPublishedCount < (strategy.postsPerWeek ?? 0)) {
-        created.push(
-          await this.opportunitiesService.createIfMissing({
-            brandId: strategyBrandId ?? '',
-            decisionReason: 'Evergreen cadence filled a weekly publishing gap.',
-            estimatedCreditCost: this.estimateOpportunityCost(
-              this.resolveFormatsForStrategy(strategy),
-            ),
-            expectedTrafficScore: 55,
-            formatCandidates: this.resolveFormatsForStrategy(strategy),
-            metadata: { trigger: 'weekly-gap' },
-            organizationId: strategyOrganizationId,
-            platformCandidates: platforms,
-            priorityScore: this.computePriorityScore(strategy, {
-              costEfficiency: 100 / this.defaultTextOpportunityCost,
-              expectedTraffic: 55,
-              freshness: 50,
-              historicalConfidence: 60,
-              relevance: this.computeTopicRelevance(strategy, defaultTopic),
-            }),
-            relevanceScore: this.computeTopicRelevance(strategy, defaultTopic),
-            sourceRef: `evergreen:${defaultTopic}`,
-            sourceType: 'evergreen',
-            strategyId,
-            topic: defaultTopic,
-          }),
-        );
-      }
-    }
-
-    return this.opportunitiesService.listOpenByStrategy(
-      strategyId,
-      strategyOrganizationId,
-    );
-  }
-
-  private selectOpportunities(
-    strategy: AgentStrategyDocument,
-    opportunities: AgentStrategyOpportunityDocument[],
-    pacing: BudgetPacingState,
-  ): AgentStrategyOpportunityDocument[] {
-    const selected: AgentStrategyOpportunityDocument[] = [];
-    let dailyRemaining = pacing.remainingDailyBudget;
-    let weeklyRemaining = pacing.remainingWeeklyBudget;
-    let monthlyRemaining = pacing.remainingMonthlyBudget;
-    let trendReserveRemaining = pacing.reserveTrendBudgetRemaining;
-    const isOverPace =
-      pacing.monthToDateCreditsUsed > pacing.expectedSpendToDate &&
-      pacing.expectedSpendToDate > 0;
-
-    for (const opportunity of opportunities.sort((a, b) => {
-      if (b.priorityScore !== a.priorityScore) {
-        return b.priorityScore - a.priorityScore;
-      }
-      if (b.expectedTrafficScore !== a.expectedTrafficScore) {
-        return b.expectedTrafficScore - a.expectedTrafficScore;
-      }
-      return a.estimatedCreditCost - b.estimatedCreditCost;
-    })) {
-      if (opportunity.status !== 'queued') {
-        continue;
-      }
-
-      if (opportunity.estimatedCreditCost > dailyRemaining) {
-        continue;
-      }
-      if (opportunity.estimatedCreditCost > weeklyRemaining) {
-        continue;
-      }
-      if (opportunity.estimatedCreditCost > monthlyRemaining) {
-        continue;
-      }
-
-      if (isOverPace && opportunity.sourceType !== 'trend') {
-        continue;
-      }
-
-      if (
-        isOverPace &&
-        opportunity.sourceType === 'trend' &&
-        opportunity.estimatedCreditCost > trendReserveRemaining
-      ) {
-        continue;
-      }
-
-      if (
-        strategy.agentType === 'video_creator' ||
-        opportunity.formatCandidates.includes('video')
-      ) {
-        continue;
-      }
-
-      selected.push(opportunity);
-      dailyRemaining -= opportunity.estimatedCreditCost;
-      weeklyRemaining -= opportunity.estimatedCreditCost;
-      monthlyRemaining -= opportunity.estimatedCreditCost;
-      if (opportunity.sourceType === 'trend') {
-        trendReserveRemaining -= opportunity.estimatedCreditCost;
-      }
-    }
-
-    return selected;
-  }
-
-  private async executeOpportunity(
-    strategy: AgentStrategyDocument,
-    opportunity: AgentStrategyOpportunityDocument,
-    userId: string,
-    defaultModel?: string,
-  ): Promise<{ contentGenerated: number; creditsUsed: number }> {
-    const strategyOrganizationId = this.strategyOrganizationId(strategy);
-    const targetPlatform = this.resolveOpportunityPlatform(
-      strategy,
-      opportunity,
-    );
-
-    await this.opportunitiesService.updateStatus(
-      this.opportunityId(opportunity),
-      strategyOrganizationId,
-      'generating',
-    );
-
-    const format = opportunity.formatCandidates[0] ?? 'text';
-
-    if (format === 'video') {
-      return this.handleVideoHold(opportunity, strategyOrganizationId);
-    }
-
-    const draft = await this.generateAutopilotDraft({
-      defaultModel,
-      format,
-      opportunity,
-      organizationId: strategyOrganizationId,
-      platform: targetPlatform,
-      strategy,
-    });
-    if (!draft) {
-      await this.opportunitiesService.updateStatus(
-        this.opportunityId(opportunity),
-        strategyOrganizationId,
-        'held',
-        { decisionReason: 'No content draft was produced.' },
-      );
-      return { contentGenerated: 0, creditsUsed: 0 };
-    }
-
-    const autopilotMetadata = await this.persistAutopilotMetadata({
-      draft,
-      format,
-      opportunity,
-      strategy,
-    });
-    const draftContent = draft.content ?? '';
-
-    const gate = await this.evaluateDraft(
-      strategy,
-      strategyOrganizationId,
-      format,
-      draftContent,
-      draft.mediaUrls?.[0],
-      targetPlatform,
-    );
-
-    if (gate.decision === 'revise' && format === 'text') {
-      const revision = await this.reviseAndReEvaluate({
-        autopilotMetadata,
-        draft,
-        draftContent,
-        format,
-        gate,
-        opportunity,
-        organizationId: strategyOrganizationId,
-        platform: targetPlatform,
-        strategy,
-        userId,
-      });
-
-      if (revision.terminal) {
-        return revision.result;
-      }
-    } else if (gate.decision !== 'approved') {
-      return this.handleGateRejection({
-        draft,
-        gate,
-        opportunity,
-        organizationId: strategyOrganizationId,
-      });
-    }
-
-    return this.finalizeApprovalAndHandoff({
-      draft,
-      draftContent,
-      format,
-      gate,
-      opportunity,
-      organizationId: strategyOrganizationId,
-      platform: targetPlatform,
-      strategy,
-      userId,
-    });
-  }
-
-  private async generateAutopilotDraft(input: {
-    defaultModel?: string;
-    format: string;
-    opportunity: AgentStrategyOpportunityDocument;
-    organizationId: string;
-    platform: string;
-    strategy: AgentStrategyDocument;
-  }): Promise<ContentDraftDocument | undefined> {
-    const generation = await this.contentGatewayService.processManualRequest(
-      input.organizationId,
-      this.strategyBrandId(input.strategy) ?? '',
-      input.format === 'image' ? 'image-generation' : 'content-writing',
-      input.format === 'image'
-        ? {
-            model: input.defaultModel,
-            prompt: this.buildImagePrompt(input.strategy, input.opportunity),
-            skillSlugs: this.strategySkillSlugs(input.strategy, [
-              'image-generation',
-            ]),
-          }
-        : {
-            model: input.defaultModel,
-            platform: input.platform,
-            skillSlugs: this.strategySkillSlugs(input.strategy, [
-              'content-writing',
-            ]),
-            topic: input.opportunity.topic,
-            variationsCount: 1,
-          },
-    );
-
-    return generation.drafts[0];
-  }
-
-  private async persistAutopilotMetadata(input: {
-    draft: ContentDraftDocument;
-    format: string;
-    opportunity: AgentStrategyOpportunityDocument;
-    strategy: AgentStrategyDocument;
-  }): Promise<Record<string, unknown>> {
-    const metadata: Record<string, unknown> = {
-      ...((input.draft.metadata ?? {}) as Record<string, unknown>),
-      autopilotFormat: input.format,
-      autopilotOpportunityId: this.opportunityId(input.opportunity),
-      autopilotSourceType: input.opportunity.sourceType,
-      autopilotStrategyId: this.strategyId(input.strategy),
-      budgetCost: input.opportunity.estimatedCreditCost,
-      goalProfile: input.strategy.goalProfile,
-    };
-
-    await this.contentDraftsService.patch(this.draftId(input.draft), {
-      metadata,
-    } as never);
-
-    return metadata;
-  }
-
-  private async handleVideoHold(
-    opportunity: AgentStrategyOpportunityDocument,
-    organizationId: string,
-  ): Promise<{ contentGenerated: number; creditsUsed: number }> {
-    await this.opportunitiesService.updateStatus(
-      this.opportunityId(opportunity),
-      organizationId,
-      'held',
-      {
-        decisionReason:
-          'Video opportunities remain draft-only until stronger evaluation and generation support lands.',
-      },
-    );
-    return { contentGenerated: 0, creditsUsed: 0 };
-  }
-
-  private async reviseAndReEvaluate(input: {
-    autopilotMetadata: Record<string, unknown>;
-    draft: ContentDraftDocument;
-    draftContent: string;
-    format: string;
-    gate: PublishGateResult;
-    opportunity: AgentStrategyOpportunityDocument;
-    organizationId: string;
-    platform: string;
-    strategy: AgentStrategyDocument;
-    userId: string;
-  }): Promise<
-    | {
-        result: { contentGenerated: number; creditsUsed: number };
-        terminal: true;
-      }
-    | { terminal: false }
-  > {
-    const draftId = this.draftId(input.draft);
-    const optimization = await this.optimizersService.optimizeContent(
-      {
-        content: input.draftContent,
-        contentType: 'caption',
-        goals: ['engagement', 'reach'],
-        platform: input.platform,
-      },
-      input.organizationId,
-      input.userId,
-    );
-
-    await this.contentDraftsService.patch(draftId, {
-      content: optimization.optimized,
-      metadata: {
-        ...input.autopilotMetadata,
-        revisionInstructions: input.gate.revisionInstructions,
-      },
-    } as never);
-
-    const revisedGate = await this.evaluateDraft(
-      input.strategy,
-      input.organizationId,
-      input.format,
-      optimization.optimized,
-      undefined,
-      input.platform,
-    );
-
-    if (revisedGate.decision !== 'approved') {
-      await this.contentDraftsService.reject(
-        draftId,
-        input.organizationId,
-        revisedGate.reasons.join(' '),
-      );
-      await this.opportunitiesService.updateStatus(
-        this.opportunityId(input.opportunity),
-        input.organizationId,
-        'discarded',
-        { decisionReason: revisedGate.reasons.join(' ') },
-      );
-      return {
-        result: {
-          contentGenerated: 1,
-          creditsUsed: input.opportunity.estimatedCreditCost,
-        },
-        terminal: true,
-      };
-    }
-
-    return { terminal: false };
-  }
-
-  private async handleGateRejection(input: {
-    draft: ContentDraftDocument;
-    gate: PublishGateResult;
-    opportunity: AgentStrategyOpportunityDocument;
-    organizationId: string;
-  }): Promise<{ contentGenerated: number; creditsUsed: number }> {
-    if (input.gate.decision === 'discard' || input.gate.decision === 'hold') {
-      await this.contentDraftsService.reject(
-        this.draftId(input.draft),
-        input.organizationId,
-        input.gate.reasons.join(' '),
-      );
-    }
-
-    await this.opportunitiesService.updateStatus(
-      this.opportunityId(input.opportunity),
-      input.organizationId,
-      input.gate.decision === 'hold' ? 'held' : 'discarded',
-      { decisionReason: input.gate.reasons.join(' ') },
-    );
-
-    return {
-      contentGenerated: 1,
-      creditsUsed: input.opportunity.estimatedCreditCost,
-    };
-  }
-
-  private async finalizeApprovalAndHandoff(
-    input: FinalizeOpportunityInput,
-  ): Promise<{ contentGenerated: number; creditsUsed: number }> {
-    if (input.format === 'text' && this.shouldAutoPublish(input.strategy)) {
-      return this.finalizeAutoPublish(input);
-    }
-
-    return this.finalizeManualReviewHandoff(input);
-  }
-
-  private async finalizeAutoPublish(
-    input: FinalizeOpportunityInput,
-  ): Promise<{ contentGenerated: number; creditsUsed: number }> {
-    const {
-      draftContent,
-      gate,
-      opportunity,
-      organizationId,
-      platform,
-      strategy,
-      userId,
-    } = input;
-    const draftId = this.draftId(input.draft);
-    const opportunityId = this.opportunityId(opportunity);
-
-    const publishResult = await this.publishTextDraft(
-      strategy,
-      draftId,
-      draftContent,
-      opportunity.platformCandidates,
-      userId,
-    );
-
-    if (publishResult.published) {
-      await this.opportunitiesService.updateStatus(
-        opportunityId,
-        organizationId,
-        'published',
-        {
-          decisionReason:
-            'Draft passed publish gate and was converted into pending posts.',
-        },
-      );
-      return {
-        contentGenerated: 1,
-        creditsUsed: opportunity.estimatedCreditCost,
-      };
-    }
-
-    await this.contentDraftsService.approve(draftId, organizationId, userId);
-
-    const reviewHandoff = await this.createPublishingInboxHandoff({
-      draftContent,
-      draftId,
-      format: this.resolveReviewBatchItemFormat(platform),
-      gate,
-      opportunity,
-      organizationId,
-      platform,
-      strategy,
-      userId,
-    });
-
-    await this.opportunitiesService.updateStatus(
-      opportunityId,
-      organizationId,
-      'approved',
-      {
-        decisionReason: reviewHandoff
-          ? `Draft passed publish gate but no connected credential was available, so it was handed off to publishing inbox batch ${reviewHandoff.batchId}.`
-          : 'Draft passed publish gate but no connected credential was available.',
-      },
-    );
-
-    return {
-      contentGenerated: 1,
-      creditsUsed: opportunity.estimatedCreditCost,
-    };
-  }
-
-  private async finalizeManualReviewHandoff(
-    input: FinalizeOpportunityInput,
-  ): Promise<{ contentGenerated: number; creditsUsed: number }> {
-    const {
-      draft,
-      draftContent,
-      format,
-      gate,
-      opportunity,
-      organizationId,
-      platform,
-      strategy,
-      userId,
-    } = input;
-    const draftId = this.draftId(draft);
-
-    await this.contentDraftsService.approve(draftId, organizationId, userId);
-
-    const reviewHandoff =
-      format === 'image'
-        ? await this.createPublishingInboxHandoff({
-            draftContent,
-            draftId,
-            format: ContentFormat.IMAGE,
-            gate,
-            mediaUrl: draft.mediaUrls?.[0],
-            opportunity,
-            organizationId,
-            platform,
-            strategy,
-            userId,
-          })
-        : format === 'text'
-          ? await this.createPublishingInboxHandoff({
-              draftContent,
-              draftId,
-              format: this.resolveReviewBatchItemFormat(platform),
-              gate,
-              opportunity,
-              organizationId,
-              platform,
-              strategy,
-              userId,
-            })
-          : null;
-
-    await this.opportunitiesService.updateStatus(
-      this.opportunityId(opportunity),
-      organizationId,
-      'approved',
-      {
-        decisionReason:
-          format === 'image'
-            ? reviewHandoff
-              ? `Image passed quality gate and was handed off to publishing inbox batch ${reviewHandoff.batchId}.`
-              : 'Image passed quality gate and was approved for downstream review/publishing.'
-            : reviewHandoff
-              ? `Draft passed publish gate and was handed off to publishing inbox batch ${reviewHandoff.batchId}.`
-              : 'Draft passed publish gate and was approved.',
-      },
-    );
-
-    return {
-      contentGenerated: 1,
-      creditsUsed: opportunity.estimatedCreditCost,
-    };
-  }
-
-  private async evaluateDraft(
-    strategy: AgentStrategyDocument,
-    organizationId: string,
-    format: string,
-    content: string,
-    mediaUrl: string | undefined,
-    platform: string,
-  ): Promise<PublishGateResult> {
-    if (format === 'image') {
-      if (!mediaUrl) {
-        return {
-          decision: 'hold',
-          overallScore: 0,
-          reasons: ['Image draft has no media URL to evaluate.'],
-          revisionInstructions: [],
-          scoreBreakdown: {},
-        };
-      }
-
-      const evaluation = (await this.evaluationsOperationsService.evaluateImage(
-        mediaUrl,
-        {
-          platform,
-          prompt: content,
-        },
-        organizationId,
-      )) as ImageEvaluationResult;
-
-      const technicalOverall = Number(
-        evaluation?.scores?.technical?.overall ?? evaluation?.overallScore ?? 0,
-      );
-      const brandOverall = Number(
-        evaluation?.scores?.brand?.overall ?? evaluation?.overallScore ?? 0,
-      );
-      const engagementOverall = Number(
-        evaluation?.scores?.engagement?.overall ??
-          evaluation?.overallScore ??
-          0,
-      );
-      const overallScore = Number(evaluation?.overallScore ?? 0);
-
-      return {
-        decision:
-          overallScore >= (strategy.publishPolicy?.minImageScore ?? 75)
-            ? 'approved'
-            : 'hold',
-        overallScore,
-        reasons:
-          overallScore >= (strategy.publishPolicy?.minImageScore ?? 75)
-            ? ['Image cleared the autopilot quality gate.']
-            : ['Image quality score did not meet the publish threshold.'],
-        revisionInstructions:
-          overallScore >= (strategy.publishPolicy?.minImageScore ?? 75)
-            ? []
-            : [
-                'Improve composition and scroll-stop strength.',
-                'Increase brand fit and reduce visible generation artifacts.',
-              ],
-        scoreBreakdown: {
-          artifactRisk: technicalOverall,
-          brandFit: brandOverall,
-          composition: technicalOverall,
-          lightingContrast: technicalOverall,
-          overlayLegibility: technicalOverall,
-          scrollStopStrength: engagementOverall,
-        },
-      };
-    }
-
-    const analysis = (await this.optimizersService.analyzeContent(
-      {
-        content,
-        contentType: 'caption',
-        goals: ['engagement', 'reach'],
-        platform,
-      },
-      organizationId,
-    )) as OptimizerAnalysisResult;
-
-    const hasCTA =
-      analysis.metadata?.hasCallToAction ??
-      /comment|click|learn more|reply|share|visit/i.test(content);
-    const overallScore = Number(analysis.overallScore ?? 0);
-    const ctaRequired = strategy.goalProfile === 'reach_traffic';
-    const minPostScore = strategy.publishPolicy?.minPostScore ?? 70;
-    const reasons: string[] = [];
-
-    if (overallScore < minPostScore) {
-      reasons.push('Post quality score fell below the publish threshold.');
-    }
-    if (ctaRequired && !hasCTA) {
-      reasons.push('Reach/traffic mode requires a visible call-to-action.');
-    }
-
-    return {
-      decision:
-        reasons.length === 0
-          ? 'approved'
-          : overallScore >= Math.max(50, minPostScore - 10)
-            ? 'revise'
-            : 'discard',
-      overallScore,
-      reasons:
-        reasons.length === 0
-          ? [
-              'Post cleared the autopilot quality gate.',
-              ...(ctaRequired && hasCTA
-                ? [
-                    'Draft includes a visible call-to-action for traffic intent.',
-                  ]
-                : []),
-            ]
-          : reasons,
-      revisionInstructions: [
-        'Strengthen the opening hook.',
-        'Improve clarity and readability.',
-        'Add a clear call-to-action aligned to traffic intent.',
-      ],
-      scoreBreakdown: {
-        clarity: Number(analysis.breakdown?.clarity ?? 0),
-        hook: Number(analysis.breakdown?.engagement ?? 0),
-        platformFit: Number(analysis.breakdown?.platformOptimization ?? 0),
-        readability: Number(analysis.breakdown?.readability ?? 0),
-      },
-    };
-  }
-
-  private async createPublishingInboxHandoff(input: {
-    draftContent: string;
-    draftId: string;
-    format: ReviewBatchItemFormat;
-    gate: PublishGateResult;
-    mediaUrl?: string;
-    opportunity: AgentStrategyOpportunityDocument;
-    organizationId: string;
-    platform?: string;
-    strategy: AgentStrategyDocument;
-    userId: string;
-  }): Promise<{
-    batchId: string;
-    postId?: string;
-    reviewItemId?: string;
-  } | null> {
-    if (input.format === ContentFormat.IMAGE && !input.mediaUrl) {
-      return null;
-    }
-
-    const batch = await this.batchGenerationService.createManualReviewBatch(
-      {
-        brandId: String(input.strategy.brand),
-        items: [
-          {
-            caption: input.draftContent,
-            format: input.format,
-            gateOverallScore: input.gate.overallScore,
-            gateReasons: input.gate.reasons,
-            label: `Autopilot ${input.format} review: ${input.opportunity.topic}`,
-            mediaUrl: input.mediaUrl,
-            opportunitySourceType: this.normalizeOpportunitySourceType(
-              input.opportunity.sourceType,
-            ),
-            opportunityTopic: input.opportunity.topic,
-            platform: input.platform,
-            prompt: input.draftContent,
-            sourceActionId: this.opportunityId(input.opportunity),
-            sourceWorkflowId: this.strategyId(input.strategy),
-            sourceWorkflowName: input.strategy.label ?? undefined,
-          },
-        ],
-      },
-      input.userId,
-      input.organizationId,
-    );
-
-    const reviewItem = batch.items[0];
-
-    // TODO: metadata nested field patch — retrieve current then spread when contentDraftsService supports it
-    await this.contentDraftsService.patch(input.draftId, {
-      metadata: {
-        reviewBatchId: batch.id,
-        reviewItemId: reviewItem?.id,
-        reviewPostId: reviewItem?.postId,
-      },
-    } as never);
-
-    if (reviewItem?.postId) {
-      await this.createPublishingInboxActivity({
-        batchId: batch.id,
-        format: input.format,
-        mediaUrl: input.mediaUrl,
-        organizationId: input.organizationId,
-        platform: input.platform,
-        postId: reviewItem.postId,
-        reviewItemId: reviewItem?.id,
-        strategy: input.strategy,
-        topic: input.opportunity.topic,
-        userId: input.userId,
-      });
-    }
-
-    return {
-      batchId: batch.id,
-      postId: reviewItem?.postId,
-      reviewItemId: reviewItem?.id,
-    };
-  }
-
-  private resolveReviewBatchItemFormat(
-    platform?: string,
-  ): ReviewBatchItemFormat {
-    const normalizedPlatform = platform?.toLowerCase();
-
-    if (
-      normalizedPlatform === 'beehiiv' ||
-      normalizedPlatform === 'substack' ||
-      normalizedPlatform === 'email'
-    ) {
-      return 'newsletter';
-    }
-
-    return 'post';
-  }
-
-  private async createPublishingInboxActivity(input: {
-    batchId: string;
-    format: ReviewBatchItemFormat;
-    mediaUrl?: string;
-    organizationId: string;
-    platform?: string;
-    postId: string;
-    reviewItemId?: string;
-    strategy: AgentStrategyDocument;
-    topic: string;
-    userId: string;
-  }): Promise<void> {
-    const href = `/posts/review?batch=${input.batchId}${
-      input.reviewItemId ? `&item=${input.reviewItemId}` : ''
-    }`;
-    const label = `Autopilot ${input.format} ready for review`;
-    const description = `${input.topic} is ready in the publishing inbox.`;
-
-    try {
-      await this.activitiesService.create({
-        brandId: String(this.strategyBrandId(input.strategy)),
-        entityId: input.postId,
-        entityModel: ActivityEntityModel.POST,
-        key: ActivityKey.POST_GENERATED,
-        organizationId: input.organizationId,
-        source: ActivitySource.POST_GENERATION,
-        userId: input.userId,
-        value: JSON.stringify({
-          batchId: input.batchId,
-          description,
-          format: input.format,
-          href,
-          label,
-          mediaUrl: input.mediaUrl,
-          platform: input.platform,
-          resultId: input.postId,
-          resultType:
-            input.format === ContentFormat.IMAGE
-              ? IngredientCategory.IMAGE
-              : input.format === ContentFormat.VIDEO
-                ? IngredientCategory.VIDEO
-                : undefined,
-          reviewItemId: input.reviewItemId,
-          sentence: description,
-          topic: input.topic,
-        }),
-      } as never);
-    } catch (error) {
-      this.logger.warn('Failed to create publishing inbox activity', {
-        batchId: input.batchId,
-        error,
-        postId: input.postId,
-        reviewItemId: input.reviewItemId,
-      });
-    }
-  }
-
-  private async publishTextDraft(
-    strategy: AgentStrategyDocument,
-    draftId: string,
-    content: string,
-    platforms: string[],
-    userId: string,
-  ): Promise<{ postIds: string[]; published: boolean }> {
-    const createdPostIds: string[] = [];
-
-    for (const platform of platforms) {
-      const credential = await this.credentialsService.findOne({
-        brandId: this.strategyBrandId(strategy) ?? String(strategy.brand),
-        isConnected: true,
-        isDeleted: false,
-        organizationId: this.strategyOrganizationId(strategy),
-        platform,
-      });
-
-      if (!credential) {
-        continue;
-      }
-
-      const post = await this.postsService.create({
-        brandId: String(this.strategyBrandId(strategy)),
-        category: PostCategory.TEXT,
-        credentialId: this.documentId(credential),
-        description: content,
-        organizationId: this.strategyOrganizationId(strategy),
-        platform: credential.platform,
-        scheduledDate: new Date(),
-        status: PostStatus.PENDING,
-        userId: userId,
-      } as never);
-
-      createdPostIds.push(this.documentId(post));
-    }
-
-    if (createdPostIds.length > 0) {
-      await this.contentDraftsService.patch(draftId, {
-        metadata: {
-          publishedPostIds: createdPostIds,
-        },
-        status: ContentDraftStatus.PUBLISHED,
-      } as never);
-      return {
-        postIds: createdPostIds,
-        published: true,
-      };
-    }
-
-    return {
-      postIds: [],
-      published: false,
-    };
-  }
-
-  private computeTopicRelevance(
-    strategy: AgentStrategyDocument,
-    topic: string,
-  ): number {
-    const loweredTopic = topic.toLowerCase();
-    const topics = strategy.topics?.map((item) => item.toLowerCase()) ?? [];
-    if (topics.length === 0) {
-      return 70;
-    }
-    if (
-      topics.some(
-        (item) => loweredTopic.includes(item) || item.includes(loweredTopic),
-      )
-    ) {
-      return 95;
-    }
-    return 60;
-  }
-
-  private computePriorityScore(
-    strategy: AgentStrategyDocument,
-    scores: {
-      costEfficiency: number;
-      expectedTraffic: number;
-      freshness: number;
-      historicalConfidence: number;
-      relevance: number;
-    },
-  ): number {
-    const ranking = strategy.rankingPolicy;
-    return Number(
-      (
-        scores.relevance * (ranking?.relevanceWeight ?? 0.3) +
-        scores.freshness * (ranking?.freshnessWeight ?? 0.2) +
-        scores.expectedTraffic * (ranking?.expectedTrafficWeight ?? 0.2) +
-        scores.historicalConfidence *
-          (ranking?.historicalConfidenceWeight ?? 0.15) +
-        scores.costEfficiency * (ranking?.costEfficiencyWeight ?? 0.15)
-      ).toFixed(2),
-    );
-  }
-
-  private estimateOpportunityCost(formats: string[]): number {
-    if (formats.includes('image')) {
-      return this.defaultImageOpportunityCost;
-    }
-    if (formats.includes('video')) {
-      return 40;
-    }
-    return this.defaultTextOpportunityCost;
-  }
-
-  private resolveFormatsForStrategy(strategy: AgentStrategyDocument): string[] {
-    if (String(strategy.agentType).includes('image')) {
-      return ['image'];
-    }
-    if (String(strategy.agentType).includes('video')) {
-      return ['video'];
-    }
-    if ((strategy.contentMix?.imagePercent ?? 0) > 50) {
-      return ['image', 'text'];
-    }
-    return ['text'];
-  }
-
-  private buildImagePrompt(
-    strategy: AgentStrategyDocument,
-    opportunity: AgentStrategyOpportunityDocument,
-  ): string {
-    return `Create an on-brand social image for ${strategy.label}. Topic: ${opportunity.topic}. Keep it high contrast, clean, scroll-stopping, and aligned to ${strategy.goalProfile}.`;
-  }
-
-  private shouldAutoPublish(strategy: AgentStrategyDocument): boolean {
-    return (
-      strategy.autonomyMode === 'auto_publish' &&
-      Boolean(strategy.publishPolicy?.autoPublishEnabled)
-    );
-  }
-
-  private resolveReportWindow(reportType: AgentStrategyReportType): {
-    periodEnd: Date;
-    periodStart: Date;
-  } {
-    const periodEnd = new Date();
-    const periodStart = new Date(periodEnd);
-    periodStart.setUTCDate(
-      periodStart.getUTCDate() - (reportType === 'weekly' ? 7 : 1),
-    );
-
-    return { periodEnd, periodStart };
   }
 }
